pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["hepitk <heino.pitkanen@helsinki.fi>"]

[tool.poetry.dependencies]
python = "^3.6"
python-docx = "^0.8.10"
invoke = "^1.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
coverage = "^5.5"
pylint = "^2.7.4"
autopep8 = "^1.5.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
